requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
mysql-connector-python>=8.1
pymysql>=1.1
requests>=2.31
python-dotenv>=1.0
plotly>=5.17
matplotlib>=3.7
seaborn>=0.12
psutil>=5.9
reportlab>=4.0
openpyxl>=3.1
pytest>=7.4
black>=23.7
flake8>=6.1